perf: remove duplicated code in spec transformer of word-cloud
