+++ packages/vchart/src/chart/word-cloud/3d/word-cloud-3d-transformer.ts
@@ -6,27 +6,9 @@ export class WordCloud3dChartSpecTransformer<
   T extends IWordCloud3dChartSpec = IWordCloud3dChartSpec
 > extends BaseWordCloudChartSpecTransformer<AdaptiveSpec<T, 'type' | 'series'>> {
   protected _getDefaultSeriesSpec(spec: IWordCloud3dChartSpec): any {
-    const series: any = super._getDefaultSeriesSpec(spec, [
-      'nameField',
-      'valueField',
-      'fontFamilyField',
-      'fontWeightField',
-      'fontStyleField',
-      'colorHexField',
-      'colorMode',
-      'colorList',
-      'rotateAngles',
-      'fontWeightRange',
-      'fontSizeRange',
-      'depth_3d',
-      'maskShape',
-      'keepAspect',
-      'random',
-      'wordCloudConfig',
-      'wordCloudShapeConfig',
-      'word',
-      'fillingWord'
-    ]);
+    const series: any = super._getDefaultSeriesSpec(spec as any);
+    series.depth_3d = spec.depth_3d;
+
     return series;
   }
 }
