+++ packages/vchart/src/chart/word-cloud/index.ts
@@ -1,4 +1,3 @@
 export * from './word-cloud';
 export * from './3d';
 export * from './interface';
-export * from './word-cloud-transformer';
